+++ endpoint/disk/disk_test.go
@@ -51,7 +51,7 @@ func TestDiskEndpoint(t *testing.T) {
 			StartTime: time.Unix(0, 0),
 			EndTime:   time.Unix(1, 0),
 			Value: metrics.MetricValue{
-				IntValue: 10,
+				Int64Value: 10,
 			},
 		},
 	})
@@ -75,7 +75,7 @@ func TestDiskEndpoint(t *testing.T) {
 			StartTime: time.Unix(2, 0),
 			EndTime:   time.Unix(3, 0),
 			Value: metrics.MetricValue{
-				IntValue: 10,
+				Int64Value: 10,
 			},
 		},
 	})
